# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Implements the Mailio server based on https://mirs.mail.io/ specifications",
        "title": "Mailio Server API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/didmessage": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Send end-to-end encrypted message to DID recipients",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messaging"
                ],
                "summary": "Send end-to-end encrypted message to DID recipients",
                "parameters": [
                    {
                        "description": "didcomm-encrypted+json",
                        "name": "handshake",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.DIDCommMessage"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/types.DIDCommApiResponse"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "401": {
                        "description": "invalid signature or unauthorized to send messages",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "429": {
                        "description": "rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/smtp": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Send SMTP email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Messaging"
                ],
                "summary": "Send SMTP email",
                "parameters": [
                    {
                        "description": "smtp email",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mailiosmtp.Mail"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/mailiosmtp.Mail"
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "401": {
                        "description": "invalid signature or unauthorized to send messages",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "429": {
                        "description": "rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    }
                }
            }
        },
...
        "/webhook/mailgun_mime": {
            "post": {
                "description": "Receive a new SMTP email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Smtp Webhook Handler"
                ],
                "summary": "Receive a new SMTP email",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "401": {
                        "description": "invalid signature/not authorized",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/api.ApiError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code is the HTTP status code",
                    "type": "integer"
                },
                "message": {
                    "description": "Message is the error message",
                    "type": "string"
                }
            }
        },
...
        "mailiosmtp.Mail": {
            "type": "object",
            "properties": {
                "attachments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mailiosmtp.SmtpAttachment"
                    }
                },
                "bcc": {
                    "description": "The email addresses of the BCC recipients.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mail.Address"
                    }
                },
                "bodyHTMLWithoutUnsafeTags": {
                    "description": "The HTML version of the email with removed unsafe tags",
                    "type": "string"
                },
                "bodyHtml": {
                    "description": "The HTML version of the email.",
                    "type": "string"
                },
                "bodyInlinePart": {
                    "description": "The raw inline content of the email.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mailiosmtp.MailBodyRaw"
                    }
                },
                "bodyText": {
                    "description": "The text version of the email.",
                    "type": "string"
                },
                "cc": {
                    "description": "The email addresses of the CC recipients.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mail.Address"
                    }
                },
                "dkimVerdict": {
                    "description": "optional, dkim verdict",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mailiosmtp.VerdictStatus"
                        }
                    ]
                },
                "dmarcVerdict": {
                    "description": "optional, dmarc verdict",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mailiosmtp.VerdictStatus"
                        }
                    ]
                },
                "from": {
                    "description": "The email address of the original sender.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mail.Address"
                        }
                    ]
                },
                "headers": {
                    "description": "The email headers. (one header can be specified multiple times with different values)",
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "messageId": {
                    "description": "message id",
                    "type": "string"
                },
                "replyTo": {
                    "description": "The email address to which bounces (undeliverable notifications) are to be forwarded.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mail.Address"
                    }
                },
                "sizeAttachmentsBytes": {
                    "description": "The size of the attachments in bytes.",
                    "type": "integer"
                },
                "sizeBytes": {
                    "description": "The size of the email in bytes.",
                    "type": "integer"
                },
                "sizeHtmlBodyBytes": {
                    "description": "The size of the HTML body in bytes.",
                    "type": "integer"
                },
                "sizeInlineBytes": {
                    "description": "The size of the inline content in bytes.",
                    "type": "integer"
                },
                "spamVerdict": {
                    "description": "optional, spam verdict",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mailiosmtp.VerdictStatus"
                        }
                    ]
                },
                "spfVerdict": {
                    "description": "optinal, spf verdict",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mailiosmtp.VerdictStatus"
                        }
                    ]
                },
                "subject": {
                    "type": "string"
                },
                "timestamp": {
                    "description": "since epoch in miliseconds",
                    "type": "integer"
                },
                "to": {
                    "description": "The email addresses of the recipients.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mail.Address"
                    }
                },
                "virusVerdict": {
                    "description": "optional, virus verdict",
                    "allOf": [
                        {
                            "$ref": "#/definitions/mailiosmtp.VerdictStatus"
                        }
                    ]
                }
            }
        },
...
        "types.DIDCommApiResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "mtpStatusCodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.MTPStatusCode"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "types.DIDCommMessage": {
            "type": "object",
            "required": [
                "from",
                "id",
                "to",
                "type"
            ],
            "properties": {
                "attachments": {
                    "description": "attachments to the message                                                // MTP status message",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.EncryptedAttachment"
                    }
                },
                "body": {
                    "description": "the body attribute contains all the data and structure defined uniquely for the schema associated with the type attribute. It MUST be a JSON object conforming to RFC 7159                              // the encrypted message body",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.EncryptedBody"
                        }
                    ]
                },
                "createdTime": {
                    "description": "time of message creation in UTC milliseconds since epoch",
                    "type": "integer"
                },
                "expiresTime": {
                    "description": "sender will abort the protocol if it doesn't get a response by this time (UTC milliseconds since epoch)",
                    "type": "integer"
                },
                "from": {
                    "description": "sender DID required because all mailio messages are encrypted",
                    "type": "string"
                },
                "fromPrior": {
                    "description": "A DID is rotated by sending a message of any type to the recipient to be notified of the rotation",
                    "type": "string"
                },
                "id": {
                    "description": "globally unique message identifier UUID (RFC 4122) recommended",
                    "type": "string"
                },
                "intent": {
                    "description": "the intent of the message (if empty, ordinary message",
                    "type": "string",
                    "enum": [
                        "message",
                        "handshake",
                        "delivery"
                    ]
                },
                "next": {
                    "description": "in case forward message",
                    "type": "string"
                },
                "plainBodyBase64": {
                    "description": "the plain text message body, base64 encoded (optional)",
                    "type": "string"
                },
                "pthid": {
                    "description": "parent thread identifier. Uniquely identifies the parent thread that the message belongs to. If not included, the message is the first message in the thread.",
                    "type": "string"
                },
                "thid": {
                    "description": "thread identifier. Uniquely identifies the thread that the message belongs to. If not included, the id property of the message MUST be treated as the value of the thid.",
                    "type": "string"
                },
                "to": {
                    "description": "in format: did:web:mail.io:0xabc -\u003e recipient DIDs",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "type": {
                    "description": "a valid message type URI (MUST be: application/didcomm-encrypted+json or application/didcomm-signed+json or application/mailio-smtp+json)",
                    "type": "string",
                    "enum": [
                        "application/didcomm-encrypted+json",
                        "application/didcomm-signed+json",
                        "application/mailio-smtp+json"
                    ]
                }
            }
        },
...
        "types.EncryptedAttachment": {
            "type": "object",
            "required": [
                "data",
                "id",
                "mediaType"
            ],
            "properties": {
                "data": {
                    "description": "the encrypted message body",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.EncryptedAttachmentData"
                        }
                    ]
                },
                "description": {
                    "description": "a human-readable description of the attachment (optional)",
                    "type": "string"
                },
                "id": {
                    "description": "a globally unique identifier for the attachment",
                    "type": "string"
                },
                "mediaType": {
                    "description": "the media type of the attachment",
                    "type": "string"
                }
            }
        },
...
        "types.EncryptedBody": {
            "type": "object",
            "required": [
                "ciphertext",
                "iv",
                "protected",
                "recipients",
                "signature",
                "tag"
            ],
            "properties": {
                "ciphertext": {
                    "description": "the encrypted message",
                    "type": "string"
                },
                "iv": {
                    "description": "the initialization vector (nonce)",
                    "type": "string"
                },
                "protected": {
                    "description": "the protected header",
                    "type": "string"
                },
                "recipients": {
                    "description": "the recipients of the message (at least one)",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/types.Recipient"
                    }
                },
                "signature": {
                    "description": "JWS digital signature",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.Signature"
                        }
                    ]
                },
                "tag": {
                    "description": "integrity check on the encrypted message",
                    "type": "string"
                }
            }
        },
...
        "types.MTPStatusCode": {
            "type": "object",
            "required": [
                "class",
                "detail",
                "subject"
            ],
            "properties": {
                "address": {
                    "description": "Address of the recipient or sender",
                    "type": "string"
                },
                "class": {
                    "description": "Represents the class of the status code (2, 4, 5)",
                    "type": "integer",
                    "enum": [
                        2,
                        4,
                        5
                    ]
                },
                "description": {
                    "description": "Human-readable message or description (optional)",
                    "type": "string"
                },
                "detail": {
                    "description": "Represents the detail of the status code",
                    "type": "integer"
                },
                "subject": {
                    "description": "Represents the subject category of the status code",
                    "type": "integer",
                    "minimum": 0
                },
                "timestamp": {
                    "description": "Unix timestamp in milliseconds",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}